[runtime-info] Move population to helper
