+++ src/daemon/daemon.cpp
@@ -18,6 +18,7 @@
 #include "daemon.h"
 #include "base_cloud_init_config.h"
 #include "instance_settings_handler.h"
+#include "runtime_instance_info_helper.h"
 #include "snapshot_settings_handler.h"
 
 #include <multipass/alias_definition.h>
@@ -1259,84 +1260,6 @@ void populate_snapshot_info(mp::VirtualMachine& vm,
         snapshot_info->add_children(child);
 
     populate_snapshot_fundamentals(snapshot, fundamentals);
-}
-
-struct RuntimeInfoKeys
-{
-public:
-    static constexpr auto loadavg_key = "loadavg";
-    static constexpr auto mem_usage_key = "mem_usage";
-    static constexpr auto mem_total_key = "mem_total";
-    static constexpr auto disk_usage_key = "disk_usage";
-    static constexpr auto disk_total_key = "disk_total";
-    static constexpr auto cpus_key = "cpus";
-    static constexpr auto current_release_key = "current_release";
-};
-
-struct RuntimeInfoCmds
-{
-private:
-    using Keys = RuntimeInfoKeys;
-    static constexpr auto key_val_cmd = R"(echo {}: \"$(eval "{}")\")";
-    static constexpr std::array key_cmds_pairs{
-        std::pair{Keys::loadavg_key, "cat /proc/loadavg | cut -d ' ' -f1-3"},
-        std::pair{Keys::mem_usage_key, R"(free -b | grep 'Mem:' | awk '{printf \$3}')"},
-        std::pair{Keys::mem_total_key, R"(free -b | grep 'Mem:' | awk '{printf \$2}')"},
-        std::pair{Keys::disk_usage_key, "df -t ext4 -t vfat --total -B1 --output=used | tail -n 1"},
-        std::pair{Keys::disk_total_key, "df -t ext4 -t vfat --total -B1 --output=size | tail -n 1"},
-        std::pair{Keys::cpus_key, "nproc"},
-        std::pair{Keys::current_release_key, R"(cat /etc/os-release | grep 'PRETTY_NAME' | cut -d \\\" -f2)"}};
-
-    inline static const std::array cmds = [] {
-        constexpr auto n = key_cmds_pairs.size();
-        std::array<std::string, key_cmds_pairs.size()> ret;
-        for (std::size_t i = 0; i < n; ++i)
-        {
-            const auto [key, cmd] = key_cmds_pairs[i];
-            ret[i] = fmt::format(key_val_cmd, key, cmd);
-        }
-
-        return ret;
-    }();
-
-public:
-    inline static const std::string sequential_composite_cmd = fmt::to_string(fmt::join(cmds, "; "));
-    inline static const std::string parallel_composite_cmd = fmt::format("{} & wait", fmt::join(cmds, "& "));
-};
-
-void populate_instance_runtime_info(mp::VirtualMachine& vm,
-                                    mp::DetailedInfoItem* info,
-                                    mp::InstanceDetails* instance_info,
-                                    const std::string& original_release,
-                                    bool parallelize)
-{
-    using Keys = RuntimeInfoKeys;
-    using Cmds = RuntimeInfoCmds;
-
-    const auto& cmd = parallelize ? Cmds::parallel_composite_cmd : Cmds::sequential_composite_cmd;
-    auto results = YAML::Load(vm.ssh_exec(cmd));
-
-    instance_info->set_load(results[Keys::loadavg_key].as<std::string>());
-    instance_info->set_memory_usage(results[Keys::mem_usage_key].as<std::string>());
-    info->set_memory_total(results[Keys::mem_total_key].as<std::string>());
-    instance_info->set_disk_usage(results[Keys::disk_usage_key].as<std::string>());
-    info->set_disk_total(results[Keys::disk_total_key].as<std::string>());
-    info->set_cpu_count(results[Keys::cpus_key].as<std::string>());
-
-    auto current_release = results[Keys::current_release_key].as<std::string>();
-    instance_info->set_current_release(!current_release.empty() ? current_release : original_release);
-
-    std::string management_ip = vm.management_ipv4();
-    auto all_ipv4 = vm.get_all_ipv4();
-
-    if (MP_UTILS.is_ipv4_valid(management_ip))
-        instance_info->add_ipv4(management_ip);
-    else if (all_ipv4.empty())
-        instance_info->add_ipv4("N/A");
-
-    for (const auto& extra_ipv4 : all_ipv4)
-        if (extra_ipv4 != management_ip)
-            instance_info->add_ipv4(extra_ipv4);
 }
 } // namespace
 
@@ -3586,7 +3509,11 @@ void mp::Daemon::populate_instance_info(VirtualMachine& vm,
     timestamp->set_nanos(created_time.time().msec() * 1'000'000);
 
     if (!no_runtime_info && MP_UTILS.is_running(present_state))
-        populate_instance_runtime_info(vm, info, instance_info, original_release, vm_specs.num_cores == 1);
+        RuntimeInstanceInfoHelper::populate_runtime_info(vm,
+                                                         info,
+                                                         instance_info,
+                                                         original_release,
+                                                         vm_specs.num_cores == 1);
 }
 
 bool mp::Daemon::is_bridged(const std::string& instance_name)
